index: widen nodeChangeOffsets to 64-bit for large pangenomes
The cumulative seed-change offset (nodeChangeOffsets) was UInt32, capping
total seed changes at 4.29B. Large pangenomes cross this — e.g. sars_8M has
an estimated ~9.5B seed changes — silently truncating offsets mod 2^32 and
corrupting placement. Widen the field to List(UInt64); the writer
accumulators and the production reader were already 64-bit, so only the
schema and the test-helper IndexData offsets/accessor needed changing.

This is a wire-incompatible change, so add an on-disk formatVersion field
and reject a stale index (formatVersion != INDEX_FORMAT_VERSION) with a
clear "rebuild" message instead of a cryptic Cap'n Proto error.

+++ src/test/helpers/test_index.hpp
@@ -30,7 +30,7 @@ class IndexData {
     std::vector<uint64_t> hashes;
     std::vector<int64_t> parentCounts;
     std::vector<int64_t> childCounts;
-    std::vector<uint32_t> offsets;  // size == numNodes + 1
+    std::vector<uint64_t> offsets;  // size == numNodes + 1
 
     size_t numNodesPlusOne() const { return offsets.size(); }
 
@@ -42,7 +42,7 @@ class IndexData {
 
     int64_t childCountAt(size_t i) const { return childCounts[i]; }
 
-    uint32_t nodeChangeOffset(size_t node) const { return offsets[node]; }
+    uint64_t nodeChangeOffset(size_t node) const { return offsets[node]; }
 };
 
 // Decompress + parse an existing .idx, returning an owned IndexData.
